sample-agents/hello-agent/requirements.txt
httpx>=0.26.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
